Move current picker
