+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/CurrentDatePicker/constants.ts
@@ -1,0 +1,6 @@
+import type { DatePickerTruncationUnit } from "../../types";
+
+export const UNIT_GROUPS: DatePickerTruncationUnit[][] = [
+  ["day", "week", "month"],
+  ["quarter", "year"],
+];
