+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/CurrentDatePicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./CurrentDatePicker";
